refactor:코드 리팩토링, 수정
