+++ package.json
@@ -27,7 +27,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.5",
-    "react-icons": "^4.8.0",
+    "react-icons": "^5.0.1",
     "react-infinite-scroller": "^1.2.6",
     "react-intersection-observer": "^9.4.3",
     "react-kakao-maps-sdk": "^1.1.24",
