+++ src/pages/Main/index.tsx
@@ -4,10 +4,9 @@ import ProductForm from '@/components/ProductForm';
 import Loading from '@/components/Loading';
 import * as S from './styles';
 import logo from '../../assets/logo.svg';
-import categoryBar from '../../assets/categoryBar.svg';
-import searchBtn from '../../assets/Search.svg';
+
 import plusImage from '../../assets/plus.png';
-import { Product } from '@/types/product';
+
 import useInfiniteScroll from '@/hooks/useInfiniteScroll';
 import { Link } from 'react-router-dom';
 import useFetchProductList from '@/hooks/useFetchProductList';
@@ -60,11 +59,11 @@ const index: React.FC = () => {
         </S.scroll>
       </S.MainContainer>
 
-      <Link to="/write">
+      {/* <Link to="/write">
         <S.Button>
           <img src={plusImage} alt="" />
         </S.Button>
-      </Link>
+      </Link> */}
       <NavBar />
     </S.MainDiv>
   );
